Struct aho_corasick::ahocorasick::StreamChunkIter

struct StreamChunkIter<'a, R, S: 'a + StateID> {
    fsm: &'a Imp<S>,
    prestate: PrefilterState,
    rdr: R,
    buf: Buffer,
    state_id: S,
    search_pos: usize,
    absolute_pos: usize,
    report_pos: usize,
    pending_match: Option<Match>,
    has_empty_match_at_end: bool,
}

An iterator over chunks in an underlying reader. Each chunk either corresponds to non-matching bytes or matching bytes, but all bytes from the underlying reader are reported in sequence. There may be an arbitrary number of non-matching chunks before seeing a matching chunk.

N.B. This does not actually implement Iterator because we need to borrow from the underlying reader. But conceptually, it's still an iterator.

Fields
fsm: &'a Imp<S>

The AC automaton.

prestate: PrefilterState

State associated with this automaton's prefilter. It is a heuristic for stopping the prefilter if it's deemed ineffective.

rdr: R

The source of bytes we read from.

buf: Buffer

A fixed size buffer. This is what we actually search. There are some invariants around the buffer's size, namely, it must be big enough to contain the longest possible match.

state_id: S

The ID of the FSM state we're currently in.

search_pos: usize

The current position at which to start the next search in buf.

absolute_pos: usize

The absolute position of search_pos, where 0 corresponds to the position of the first byte read from rdr.

report_pos: usize

The ending position of the last StreamChunk that was returned to the caller. This position is used to determine whether we need to emit non-matching bytes before emitting a match.

pending_match: Option<Match>

A match that should be reported on the next call.

has_empty_match_at_end: bool

Enabled only when the automaton can match the empty string. When enabled, we need to execute one final search after consuming the reader to find the trailing empty match.

Implementations

impl<'a, R: Read, S: StateID> StreamChunkIter<'a, R, S>

fn new(ac: &'a AhoCorasick<S>, rdr: R) -> StreamChunkIter<'a, R, S>

fn next<'r>(&'r mut self) -> Option<Result<StreamChunk<'r>>>

fn unreported(&self) -> Option<usize>