Module aho_corasick::autiter

Structs

Match
Records a match in the search text.

Matches
An iterator of non-overlapping matches for in-memory text.

MatchesOverlapping
An iterator of overlapping matches for in-memory text.

StreamMatches
An iterator of non-overlapping matches for streaming text.

StreamMatchesOverlapping
An iterator of overlapping matches for streaming text.

Traits

Automaton
An abstraction over automatons and their corresponding iterators. The type parameter P is the type of the pattern that was used to construct this Automaton.

Functions

skip1

skip2

skip3

skip_to_match

step_to_match