aho_corasick

Module automaton

Provides Automaton trait for abstracting over Aho-Corasick automata.

The Automaton trait provides a way to write generic code over any Aho-Corasick automaton. It also provides access to lower level APIs that permit walking the state transitions of an Aho-Corasick automaton manually.

Modules

private 🔒

We seal the Automaton trait for now. It’s a big trait, and it’s conceivable that I might want to add new required methods, and sealing the trait permits doing that in a backwards compatible fashion. On other the hand, if you have a solid use case for implementing the trait yourself, please file an issue and we can discuss it. This was mostly done as a conservative step.

Structs

FindIter

An iterator of non-overlapping matches in a particular haystack.

FindOverlappingIter

An iterator of overlapping matches in a particular haystack.

OverlappingState

Represents the current state of an overlapping search.

Prefilter

A prefilter for accelerating a search.

StateID

The identifier of a finite automaton state.

StateIDError

This error occurs when an ID could not be constructed.

StreamChunkIter 🔒

An iterator that reports matches in a stream.

StreamFindIter

An iterator that reports matches in a stream.

Enums

Candidate

A candidate is the result of running a prefilter on a haystack at a particular position.

StreamChunk 🔒

A single chunk yielded by the stream chunk iterator.

Traits

Automaton

A trait that abstracts over Aho-Corasick automata.

Functions

fmt_state_indicator 🔒

Write a prefix “state” indicator for fmt::Debug impls. It always writes exactly two printable bytes to the given formatter.

get_match 🔒

sparse_transitions 🔒

Return an iterator of transitions in a sparse format given an iterator of all explicitly defined transitions. The iterator yields ranges of transitions, such that any adjacent transitions mapped to the same state are combined into a single range.

try_find_fwd 🔒

try_find_fwd_imp 🔒

try_find_overlapping_fwd 🔒

try_find_overlapping_fwd_imp 🔒