aho_corasick::automaton
Function fmt_state_indicator

pub(crate) fn fmt_state_indicator<A: Automaton>(
    f: &mut Formatter<'_>,
    aut: A,
    id: StateID,
) -> Result

Write a prefix “state” indicator for fmt::Debug impls. It always writes exactly two printable bytes to the given formatter.

Specifically, this tries to succinctly distinguish the different types of states: dead states, start states and match states. It even accounts for the possible overlappings of different state types. (The only possible overlapping is that of match and start states.)