aho_corasick::automaton

Function sparse_transitions

pub(crate) fn sparse_transitions<'a>(
    it: impl Iterator<Item = (u8, StateID)> + 'a,
) -> impl Iterator<Item = (u8, u8, StateID)> + 'a

Return an iterator of transitions in a sparse format given an iterator of all explicitly defined transitions. The iterator yields ranges of transitions, such that any adjacent transitions mapped to the same state are combined into a single range.