Struct aho_corasick::automaton::OverlappingState

pub struct OverlappingState { /* private fields */ }

Represents the current state of an overlapping search.

This is used for overlapping searches since they need to know something about the previous search. For example, when multiple patterns match at the same position, this state tracks the last reported pattern so that the next search knows whether to report another matching pattern or continue with the search at the next position. Additionally, it also tracks which state the last search call terminated in and the current offset of the search in the haystack.

This type provides limited introspection capabilities. The only thing a caller can do is construct it and pass it around to permit search routines to use it to track state, and to ask whether a match has been found.

Callers should always provide a fresh state constructed via OverlappingState::start when starting a new search. That same state should be reused for subsequent searches on the same Input. The state given will advance through the haystack itself. Callers can detect the end of a search when neither an error nor a match is returned.

Example

This example shows how to manually iterate over all overlapping matches. If you need this, you might consider using AhoCorasick::find_overlapping_iter instead, but this shows how to correctly use an OverlappingState.

use aho_corasick::{
    automaton::OverlappingState,
    AhoCorasick, Input, Match,
};

let patterns = &["append", "appendage", "app"];
let haystack = "append the app to the appendage";

let ac = AhoCorasick::new(patterns).unwrap();
let mut state = OverlappingState::start();
let mut matches = vec![];

loop {
    ac.find_overlapping(haystack, &mut state);
    let mat = match state.get_match() {
        None => break,
        Some(mat) => mat,
    };
    matches.push(mat);
}
let expected = vec![
    Match::must(2, 0..3),
    Match::must(0, 0..6),
    Match::must(2, 11..14),
    Match::must(2, 22..25),
    Match::must(0, 22..28),
    Match::must(1, 22..31),
];
assert_eq!(expected, matches);

Implementations

impl OverlappingState

pub fn start() -> OverlappingState

Create a new overlapping state that begins at the start state.

pub fn get_match(&self) -> Option<Match>

Return the match result of the most recent search to execute with this state.

Every search will clear this result automatically, such that if no match is found, this will always correctly report None.