Function aho_corasick::dfa::nfa_next_state_memoized

fn nfa_next_state_memoized<S: StateID>(
    nfa: &NFA<S>, 
    dfa: &Repr<S>, 
    populating: S, 
    current: S, 
    input: u8
) -> S

This returns the next NFA transition (including resolving failure transitions), except once it sees a state id less than the id of the DFA state that is currently being populated, then we no longer need to follow failure transitions and can instead query the pre-computed state id from the DFA itself.

In general, this should only be called when a failure transition is seen.