Function aho_corasick::dfa::sparse_iter

fn sparse_iter<F: FnMut(u8, u8, StateID)>(
    nnfa: &NFA,
    oldsid: StateID,
    classes: &ByteClasses,
    f: F,
)

Iterate over all possible equivalence class transitions in this state. The closure is called for all transitions with a distinct equivalence class, even those not explicitly represented in this sparse state. For any implicitly defined transitions, the given closure is called with the fail state ID.

The closure is guaranteed to be called precisely byte_classes.alphabet_len() times, once for every possible class in ascending order.