Module aho_corasick::nfa::contiguous

Provides a contiguous NFA implementation of Aho-Corasick.

This is a low-level API that generally only needs to be used in niche circumstances. When possible, prefer using AhoCorasick instead of a contiguous NFA directly. Using an NFA directly is typically only necessary when one needs access to the Automaton trait implementation.

Structs

Builder

A builder for configuring an Aho-Corasick contiguous NFA.

NFA

A contiguous NFA implementation of Aho-Corasick.