aho_corasick::nfa
Module noncontiguous

Provides a noncontiguous NFA implementation of Aho-Corasick.

This is a low-level API that generally only needs to be used in niche circumstances. When possible, prefer using AhoCorasick instead of a noncontiguous NFA directly. Using an NFA directly is typically only necessary when one needs access to the Automaton trait implementation.

Structs

Builder

A builder for configuring an Aho-Corasick noncontiguous NFA.

NFA

A noncontiguous NFA implementation of Aho-Corasick.