Struct aho_corasick::nfa::noncontiguous::Builder

pub struct Builder { /* private fields */ }

A builder for configuring an Aho-Corasick noncontiguous NFA.

This builder has a subset of the options available to a AhoCorasickBuilder. Of the shared options, their behavior is identical.

Implementations

impl Builder

pub fn new() -> Builder

Create a new builder for configuring an Aho-Corasick noncontiguous NFA.

pub fn build<I, P>(&self, patterns: I) -> Result<NFA, BuildError>where
I: IntoIterator<Item = P>,
P: AsRef<[u8]>,

Build an Aho-Corasick noncontiguous NFA from the given iterator of patterns.

A builder may be reused to create more NFAs.

pub fn match_kind(&mut self, kind: MatchKind) -> &mut Builder

Set the desired match semantics.

See AhoCorasickBuilder::match_kind for more documentation and examples.

pub fn ascii_case_insensitive(&mut self, yes: bool) -> &mut Builder

Enable ASCII-aware case insensitive matching.

See AhoCorasickBuilder::ascii_case_insensitive for more documentation and examples.

pub fn dense_depth(&mut self, depth: usize) -> &mut Builder

Set the limit on how many states use a dense representation for their transitions. Other states will generally use a sparse representation.

See AhoCorasickBuilder::dense_depth for more documentation and examples.

pub fn prefilter(&mut self, yes: bool) -> &mut Builder

Enable heuristic prefilter optimizations.

See AhoCorasickBuilder::prefilter for more documentation and examples.