Struct aho_corasick::nfa::Compiler

struct Compiler<'a, S: StateID> {
    builder: &'a Builder,
    prefilter: Builder,
    nfa: NFA<S>,
    byte_classes: ByteClassBuilder,
}

A compiler uses a builder configuration and builds up the NFA formulation of an Aho-Corasick automaton. This roughly corresponds to the standard formulation described in textbooks.

Fields
builder: &'a Builderprefilter: Buildernfa: NFA<S>byte_classes: ByteClassBuilder
Implementations

impl<'a, S: StateID> Compiler<'a, S>

fn new(builder: &'a Builder) -> Result<Compiler<'a, S>, Error>

fn compile<I, P>(self, patterns: I) -> Result<NFA<S>, Error> where
I: IntoIterator<Item = P>,
P: AsRef<[u8]>,

fn build_trie<I, P>(&mut self, patterns: I) -> Result<(), Error> where
I: IntoIterator<Item = P>,
P: AsRef<[u8]>,

This sets up the initial prefix trie that makes up the Aho-Corasick automaton. Effectively, it creates the basic structure of the automaton, where every pattern given has a path from the start state to the end of the pattern.

fn fill_failure_transitions_standard(&mut self)

This routine creates failure transitions according to the standard textbook formulation of the Aho-Corasick algorithm.

Building failure transitions is the most interesting part of building the Aho-Corasick automaton, because they are what allow searches to be performed in linear time. Specifically, a failure transition is a single transition associated with each state that points back to the longest proper suffix of the pattern being searched. The failure transition is followed whenever there exists no transition on the current state for the current input byte. If there is no other proper suffix, then the failure transition points back to the starting state.

For example, let’s say we built an Aho-Corasick automaton with the following patterns: ‘abcd’ and ‘cef’. The trie looks like this:

         a - S1 - b - S2 - c - S3 - d - S4*
        /
    S0 - c - S5 - e - S6 - f - S7*

At this point, it should be fairly straight-forward to see how this trie can be used in a simplistic way. At any given position in the text we’re searching (called the “subject” string), all we need to do is follow the transitions in the trie by consuming one transition for each byte in the subject string. If we reach a match state, then we can report that location as a match.

The trick comes when searching a subject string like ‘abcef’. We’ll initially follow the transition from S0 to S1 and wind up in S3 after observng the ‘c’ byte. At this point, the next byte is ‘e’ but state S3 has no transition for ‘e’, so the search fails. We then would need to restart the search at the next position in ‘abcef’, which corresponds to ‘b’. The match would fail, but the next search starting at ‘c’ would finally succeed. The problem with this approach is that we wind up searching the subject string potentially many times. In effect, this makes the algorithm have worst case O(n * m) complexity, where n ~ len(subject) and m ~ len(all patterns). We would instead like to achieve a O(n + m) worst case complexity.

This is where failure transitions come in. Instead of dying at S3 in the first search, the automaton can instruct the search to move to another part of the automaton that corresponds to a suffix of what we’ve seen so far. Recall that we’ve seen ‘abc’ in the subject string, and the automaton does indeed have a non-empty suffix, ‘c’, that could potentially lead to another match. Thus, the actual Aho-Corasick automaton for our patterns in this case looks like this:

         a - S1 - b - S2 - c - S3 - d - S4*
        /                      /
       /       ----------------
      /       /
    S0 - c - S5 - e - S6 - f - S7*

That is, we have a failure transition from S3 to S5, which is followed exactly in cases when we are in state S3 but see any byte other than ‘d’ (that is, we’ve “failed” to find a match in this portion of our trie). We know we can transition back to S5 because we’ve already seen a ‘c’ byte, so we don’t need to re-scan it. We can then pick back up with the search starting at S5 and complete our match.

Adding failure transitions to a trie is fairly simple, but subtle. The key issue is that you might have multiple failure transition that you need to follow. For example, look at the trie for the patterns ‘abcd’, ‘b’, ‘bcd’ and ‘cd’:

       - a - S1 - b - S2 - c - S3 - d - S4*
      /
    S0 - b - S5* - c - S6 - d - S7*
      \
       - c - S8 - d - S9*

The failure transitions for this trie are defined from S2 to S5, S3 to S6 and S6 to S8. Moreover, state S2 needs to track that it corresponds to a match, since its failure transition to S5 is itself a match state.

Perhaps simplest way to think about adding these failure transitions is recursively. That is, if you know the failure transitions for every possible previous state that could be visited (e.g., when computing the failure transition for S3, you already know the failure transitions for S0, S1 and S2), then you can simply follow the failure transition of the previous state and check whether the incoming transition is defined after following the failure transition.

For example, when determining the failure state for S3, by our assumptions, we already know that there is a failure transition from S2 (the previous state) to S5. So we follow that transition and check whether the transition connecting S2 to S3 is defined. Indeed, it is, as there is a transition from S5 to S6 for the byte ‘c’. If no such transition existed, we could keep following the failure transitions until we reach the start state, which is the failure transition for every state that has no corresponding proper suffix.

We don’t actually use recursion to implement this, but instead, use a breadth first search of the automaton. Our base case is the start state, whose failure transition is just a transition to itself.

fn fill_failure_transitions_leftmost(&mut self)

This routine is just like fill_failure_transitions_standard, except it adds failure transitions in a way that preserves leftmost match semantics (for both leftmost-first and leftmost-longest).

The algorithms are so similar that it would be possible to write it generically. But doing so without overhead would require a bit of ceremony, so we just copy it and add in the extra leftmost logic. Moreover, the standard algorithm above is so simple that it feels like a crime to disturb it.

In effect, this proceeds just like the standard approach, but we specifically add only a subset of all failure transitions. Namely, we only add failure transitions that either do not occur after a match or failure transitions that do occur after a match but preserve the match. The comments in the implementation below should help.

N.B. The only differences in the automaton between leftmost-first and leftmost-longest are in trie construction. Otherwise, both have exactly the same set of failure transitions. leftmost-longest adds everything to the trie, where as leftmost-first skips any patterns for which there exists a prefix of it that was added earlier.

N.B. I came up with this algorithm on my own, and after scouring all of the other AC implementations I know of (Perl, Snort, many on GitHub). I couldn’t find any that implement leftmost semantics like this. Perl of course needs leftmost-first semantics, but they implement it with a seeming hack at search time instead of encoding it into the automaton. There are also a couple Java libraries that support leftmost longest semantics, but they do it by building a queue of matches at search time, which is even worse than what Perl is doing. —AG

fn queued_set(&self) -> QueuedSet<S>

Returns a set that tracked queued states.

This is only necessary when ASCII case insensitivity is enabled, since it is the only way to visit the same state twice. Otherwise, this returns an inert set that nevers adds anything and always reports false for every member test.

fn add_start_state_loop(&mut self)

Set the failure transitions on the start state to loop back to the start state. This effectively permits the Aho-Corasick automaton to match at any position. This is also required for finding the next state to terminate, namely, finding the next state should never return a fail_id.

This must be done after building the initial trie, since trie construction depends on transitions to fail_id to determine whether a state already exists or not.

fn close_start_state_loop(&mut self)

Remove the start state loop by rewriting any transitions on the start state back to the start state with transitions to the dead state.

The loop is only closed when two conditions are met: the start state is a match state and the match kind is leftmost-first or leftmost-longest. (Alternatively, if this is an anchored automaton, then the start state is always closed, regardless of aforementioned conditions.)

The reason for this is that under leftmost semantics, a start state that is also a match implies that we should never restart the search process. We allow normal transitions out of the start state, but if none exist, we transition to the dead state, which signals that searching should stop.

fn add_dead_state_loop(&mut self)

Sets all transitions on the dead state to point back to the dead state. Normally, missing transitions map back to the failure state, but the point of the dead state is to act as a sink that can never be escaped.

fn calculate_size(&mut self)

Computes the total amount of heap used by this NFA in bytes.

fn add_state(&mut self, depth: usize) -> Result<S, Error>

Add a new state to the underlying NFA with the given depth. The depth is used to determine how to represent the transitions.

If adding the new state would overflow the chosen state ID representation, then this returns an error.

fn match_kind(&self) -> MatchKind

Returns the match kind configured on the underlying builder.