Struct aho_corasick::nfa::NFA

pub struct NFA<S> {
    match_kind: MatchKind,
    start_id: S,
    max_pattern_len: usize,
    pattern_count: usize,
    heap_bytes: usize,
    prefilter: Option<PrefilterObj>,
    anchored: bool,
    byte_classes: ByteClasses,
    states: Vec<State<S>>,
}

An Aho-Corasick automaton, represented as an NFA.

This is the classical formulation of Aho-Corasick, which involves building up a prefix trie of a given set of patterns, and then wiring up failure transitions between states in order to guarantee linear time matching. The standard formulation is, technically, an NFA because of these failure transitions. That is, one can see them as enabling the automaton to be in multiple states at once. Indeed, during search, it is possible to check the transitions on multiple states for a single input byte.

This particular implementation not only supports the standard style of matching, but also provides a mode for choosing leftmost-first or leftmost-longest match semantics. When a leftmost mode is chosen, some failure transitions that would otherwise be added are elided. See the documentation of MatchKind for more details and examples on how the match semantics may differ.

If one wants a DFA, then it is necessary to first build an NFA and convert it into a DFA. Note, however, that because we’ve constrained ourselves to matching literal patterns, this does not need to use subset construction for determinization. Instead, the DFA has at most a number of states equivalent to the number of NFA states. The only real difference between them is that all failure transitions are followed and pre-computed. This uses much more memory, but also executes searches more quickly.

Fields
match_kind: MatchKind

The match semantics built into this NFA.

start_id: S

The start state id as an index into states.

max_pattern_len: usize

The length, in bytes, of the longest pattern in this automaton. This information is useful for keeping correct buffer sizes when searching on streams.

pattern_count: usize

The total number of patterns added to this automaton, including patterns that may never be matched.

heap_bytes: usize

The number of bytes of heap used by this NFA’s transition table.

prefilter: Option<PrefilterObj>

A prefilter for quickly skipping to candidate matches, if pertinent.

anchored: bool

Whether this automaton anchors all matches to the start of input.

byte_classes: ByteClasses

A set of equivalence classes in terms of bytes. We compute this while building the NFA, but don’t use it in the NFA’s states. Instead, we use this for building the DFA. We store it on the NFA since it’s easy to compute while visiting the the patterns.

states: Vec<State<S>>

A set of states. Each state defines its own transitions, a fail transition and a set of indices corresponding to matches.

The first state is always the fail state, which is used only as a sentinel. Namely, in the final NFA, no transition into the fail state exists. (Well, they do, but they aren’t followed. Instead, the state’s failure transition is followed.)

The second state (index 1) is always the dead state. Dead states are in every automaton, but only used when leftmost-{first,longest} match semantics are enabled. Specifically, they instruct search to stop at specific points in order to report the correct match location. In the standard Aho-Corasick construction, there are no transitions to the dead state.

The third state (index 2) is generally intended to be the starting or “root” state.

Implementations

impl<S: StateID> NFA<S>

pub fn byte_classes(&self) -> &ByteClasses

Returns the equivalence classes of bytes found while constructing this NFA.

Note that the NFA doesn’t actually make use of these equivalence classes. Instead, these are useful for building the DFA when desired.

pub fn prefilter_obj(&self) -> Option<&PrefilterObj>

Returns a prefilter, if one exists.

pub fn heap_bytes(&self) -> usize

Returns the total number of heap bytes used by this NFA’s transition table.

pub fn max_pattern_len(&self) -> usize

Return the length of the longest pattern in this automaton.

pub fn pattern_count(&self) -> usize

Return the total number of patterns added to this automaton.

pub fn state_len(&self) -> usize

Returns the total number of states in this NFA.

pub fn matches(&self, id: S) -> &[(usize, usize)]

Returns the matches for the given state.

pub fn iter_all_transitions<F: FnMut(u8, S)>(
&self,
byte_classes: &ByteClasses,
id: S,
f: F
)

Returns an iterator over all transitions in the given state according to the given equivalence classes, including transitions to fail_id(). The number of transitions returned is always equivalent to the number of equivalence classes.

pub fn failure_transition(&self, id: S) -> S

Returns the failure transition for the given state.

pub fn next_state(&self, current: S, input: u8) -> S

Returns the next state for the given state and input byte.

Note that this does not follow failure transitions. As such, the id returned may be fail_id.

fn state(&self, id: S) -> &State<S>

fn state_mut(&mut self, id: S) -> &mut State<S>

fn start(&self) -> &State<S>

fn start_mut(&mut self) -> &mut State<S>

fn iter_transitions_mut(&mut self, id: S) -> IterTransitionsMut<'_, S>

fn copy_matches(&mut self, src: S, dst: S)

fn copy_empty_matches(&mut self, dst: S)

fn add_dense_state(&mut self, depth: usize) -> Result<S, Error>

fn add_sparse_state(&mut self, depth: usize) -> Result<S, Error>