aho_corasick::packed
Module pattern

Structs

Pattern 🔒

A pattern that is used in packed searching.

PatternIter 🔒

An iterator over the patterns in the Patterns collection.

Patterns 🔒

A non-empty collection of non-empty patterns to search for.

Functions

is_equal_raw 🔒 ⚠

Compare n bytes at the given pointers for equality.

is_prefix 🔒

Returns true if and only if needle is a prefix of haystack.