Type Definition aho_corasick::packed::pattern::PatternID

type PatternID = u16;

The type used for representing a pattern identifier.

We don't use usize here because our packed searchers don't scale to huge numbers of patterns, so we keep things a bit smaller.