aho_corasick::packed::teddy
Module builder

Modules

x86_64 🔒

Structs

Builder 🔒

A builder for constructing a Teddy matcher.

Searcher 🔒

A searcher that dispatches to one of several possible Teddy variants.

Traits

SearcherT 🔒

A trait that provides dynamic dispatch over the different possible Teddy variants on the same algorithm.