Struct aho_corasick::packed::teddy::generic::Slim

pub(crate) struct Slim<V, const BYTES: usize> {
    teddy: Teddy<8>,
    masks: [Mask<V>; BYTES],
}

A “slim” Teddy implementation that is generic over both the vector type and the minimum length of the patterns being searched for.

Only 1, 2, 3 and 4 bytes are supported as minimum lengths.

Fields
teddy: Teddy<8>

A generic data structure for doing “slim” Teddy verification.

masks: [Mask<V>; BYTES]

The masks used as inputs to the shuffle operation to generate candidates (which are fed into the verification routines).

Implementations

impl<V: Vector, const BYTES: usize> Slim<V, BYTES>

pub(crate) unsafe fn new(patterns: Arc<Patterns>) -> Slim<V, BYTES>

Create a new “slim” Teddy searcher for the given patterns.

Panics

This panics when BYTES is any value other than 1, 2, 3 or 4.

Safety

Callers must ensure that this is okay to call in the current target for the current CPU.

pub(crate) fn memory_usage(&self) -> usize

Returns the approximate total amount of heap used by this type, in units of bytes.

pub(crate) fn minimum_len(&self) -> usize

Returns the minimum length, in bytes, that a haystack must be in order to use it with this searcher.

impl<V: Vector> Slim<V, 1>

pub(crate) unsafe fn find( &self, start: *const u8, end: *const u8, ) -> Option<Match>

Look for an occurrences of the patterns in this finder in the haystack given by the start and end pointers.

If no match could be found, then None is returned.

Safety

The given pointers representing the haystack must be valid to read from. They must also point to a region of memory that is at least the minimum length required by this searcher.

Callers must ensure that this is okay to call in the current target for the current CPU.

unsafe fn find_one(&self, cur: *const u8, end: *const u8) -> Option<Match>

Look for a match starting at the V::BYTES at and after cur. If there isn’t one, then None is returned.

Safety

The given pointers representing the haystack must be valid to read from. They must also point to a region of memory that is at least the minimum length required by this searcher.

Callers must ensure that this is okay to call in the current target for the current CPU.

unsafe fn candidate(&self, cur: *const u8) -> V

Look for a candidate match (represented as a vector) starting at the V::BYTES at and after cur. If there isn’t one, then a vector with all bits set to zero is returned.

Safety

The given pointer representing the haystack must be valid to read from.

Callers must ensure that this is okay to call in the current target for the current CPU.

impl<V: Vector> Slim<V, 2>

pub(crate) unsafe fn find( &self, start: *const u8, end: *const u8, ) -> Option<Match>

See Slim<V, 1>::find.

unsafe fn find_one( &self, cur: *const u8, end: *const u8, prev0: &mut V, ) -> Option<Match>

See Slim<V, 1>::find_one.

unsafe fn candidate(&self, cur: *const u8, prev0: &mut V) -> V

See Slim<V, 1>::candidate.

impl<V: Vector> Slim<V, 3>

pub(crate) unsafe fn find( &self, start: *const u8, end: *const u8, ) -> Option<Match>

See Slim<V, 1>::find.

unsafe fn find_one( &self, cur: *const u8, end: *const u8, prev0: &mut V, prev1: &mut V, ) -> Option<Match>

See Slim<V, 1>::find_one.

unsafe fn candidate(&self, cur: *const u8, prev0: &mut V, prev1: &mut V) -> V

See Slim<V, 1>::candidate.

impl<V: Vector> Slim<V, 4>

pub(crate) unsafe fn find( &self, start: *const u8, end: *const u8, ) -> Option<Match>

See Slim<V, 1>::find.

unsafe fn find_one( &self, cur: *const u8, end: *const u8, prev0: &mut V, prev1: &mut V, prev2: &mut V, ) -> Option<Match>

See Slim<V, 1>::find_one.

unsafe fn candidate( &self, cur: *const u8, prev0: &mut V, prev1: &mut V, prev2: &mut V, ) -> V

See Slim<V, 1>::candidate.