Module aho_corasick::packed::teddy::runtime

Structs

Mask128
A 128-bit mask for the low and high nybbles in a set of patterns. Each lane j corresponds to a bitset where the ith bit is set if and only if the nybble j is in the bucket i at a particular position.

Mask256
A 256-bit mask for the low and high nybbles in a set of patterns. Each lane j corresponds to a bitset where the ith bit is set if and only if the nybble j is in the bucket i at a particular position.

Teddy
The Teddy runtime.

TeddyFat1Mask256

TeddyFat2Mask256

TeddyFat3Mask256

TeddySlim1Mask128

TeddySlim1Mask256

TeddySlim2Mask128

TeddySlim2Mask256

TeddySlim3Mask128

TeddySlim3Mask256

Enums

Exec
Exec represents the different search strategies supported by the Teddy runtime.

Functions

members1m128⚠	ssse3
Return a candidate for Slim 128-bit Teddy, where chunk corresponds to a 16-byte window of the haystack (where the least significant byte corresponds to the start of the window), and mask1 corresponds to a low/high mask for the first byte of all patterns that are being searched.

members1m256⚠	avx2
Return a candidate for Slim 256-bit Teddy, where chunk corresponds to a 32-byte window of the haystack (where the least significant byte corresponds to the start of the window), and mask1 corresponds to a low/high mask for the first byte of all patterns that are being searched.

members2m128⚠	ssse3
Return candidates for Slim 128-bit Teddy, where chunk corresponds to a 16-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first and second bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first and second bytes in the patterns represented by the masks.

members2m256⚠	avx2
Return candidates for Slim 256-bit Teddy, where chunk corresponds to a 32-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first and second bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first and second bytes in the patterns represented by the masks.

members3m128⚠	ssse3
Return candidates for Slim 128-bit Teddy, where chunk corresponds to a 16-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first, second and third bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first, second and third bytes in the patterns represented by the masks.

members3m256⚠	avx2
Return candidates for Slim 256-bit Teddy, where chunk corresponds to a 32-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first, second and third bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first, second and third bytes in the patterns represented by the masks.