Function aho_corasick::packed::teddy::runtime::members2m128

unsafe fn members2m128(
    chunk: __m128i, 
    mask1: Mask128, 
    mask2: Mask128
) -> (__m128i, __m128i)

This is supported with target feature ssse3 only.

Return candidates for Slim 128-bit Teddy, where chunk corresponds to a 16-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first and second bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first and second bytes in the patterns represented by the masks.