Function aho_corasick::packed::teddy::runtime::members3m256

unsafe fn members3m256(
    chunk: __m256i, 
    mask1: Mask256, 
    mask2: Mask256, 
    mask3: Mask256
) -> (__m256i, __m256i, __m256i)

This is supported with target feature avx2 only.

Return candidates for Slim 256-bit Teddy, where chunk corresponds to a 32-byte window of the haystack (where the least significant byte corresponds to the start of the window), and the masks correspond to a low/high mask for the first, second and third bytes of all patterns that are being searched. The vectors returned correspond to candidates for the first, second and third bytes in the patterns represented by the masks.

Note that this can also be used for Fat Teddy, where the high 128 bits in chunk is the same as the low 128 bits, which corresponds to a 16 byte window in the haystack.