aho_corasick::packed

Module vector

Modules

x86_64_avx2 🔒

x86_64_ssse3 🔒

Traits

FatVector 🔒

This trait extends the Vector trait with additional operations to support Fat Teddy.

Vector 🔒

A trait for describing vector operations used by vectorized searchers.