Struct aho_corasick::prefilter::PrefilterState

pub struct PrefilterState {
    skips: usize,
    skipped: usize,
    max_match_len: usize,
    inert: bool,
    last_scan_at: usize,
}

PrefilterState tracks state associated with the effectiveness of a prefilter. It is used to track how many bytes, on average, are skipped by the prefilter. If this average dips below a certain threshold over time, then the state renders the prefilter inert and stops using it.

A prefilter state should be created for each search. (Where creating an iterator via, e.g., find_iter, is treated as a single search.)

Fields
skips: usize

The number of skips that has been executed.

skipped: usize

The total number of bytes that have been skipped.

max_match_len: usize

The maximum length of a match. This is used to help determine how many bytes on average should be skipped in order for a prefilter to be effective.

inert: bool

Once this heuristic has been deemed permanently ineffective, it will be inert throughout the rest of its lifetime. This serves as a cheap way to check inertness.

last_scan_at: usize

The last (absolute) position at which a prefilter scanned to. Prefilters can use this position to determine whether to re-scan or not.

Unlike other things that impact effectiveness, this is a fleeting condition. That is, a prefilter can be considered ineffective if it is at a position before last_scan_at, but can become effective again once the search moves past last_scan_at.

The utility of this is to both avoid additional overhead from calling the prefilter and to avoid quadratic behavior. This ensures that a prefilter will scan any particular byte at most once. (Note that some prefilters, like the start-byte prefilter, do not need to use this field at all, since it only looks for starting bytes.)

Implementations

impl PrefilterState

const MIN_SKIPS: usize

The minimum number of skip attempts to try before considering whether a prefilter is effective or not.

const MIN_AVG_FACTOR: usize

The minimum amount of bytes that skipping must average, expressed as a factor of the multiple of the length of a possible match.

That is, after MIN_SKIPS have occurred, if the average number of bytes skipped ever falls below MIN_AVG_FACTOR * max-match-length, then the prefilter outed to be rendered inert.

pub fn new(max_match_len: usize) -> PrefilterState

Create a fresh prefilter state.

pub fn disabled() -> PrefilterState

Create a prefilter state that always disables the prefilter.

fn update_skipped_bytes(&mut self, skipped: usize)

Update this state with the number of bytes skipped on the last invocation of the prefilter.

fn update_at(&mut self, at: usize)

Updates the position at which the last scan stopped. This may be greater than the position of the last candidate reported. For example, searching for the “rare” byte z in abczdef for the pattern abcz will report a candidate at position 0, but the end of its last scan will be at position 3.

This position factors into the effectiveness of this prefilter. If the current position is less than the last position at which a scan ended, then the prefilter should not be re-run until the search moves past that position.

pub fn is_effective(&mut self, at: usize) -> bool

Return true if and only if this state indicates that a prefilter is still effective.

The given pos should correspond to the current starting position of the search.