Module aho_corasick::state_id

Modules

private

Traits

StateID

A trait describing the representation of an automaton’s state identifier.

Functions

dead_id

Return the unique identifier for an automaton’s fail state in the chosen representation indicated by S.

fail_id

Return the unique identifier for an automaton’s fail state in the chosen representation indicated by S.

premultiply_overflow_error

Check that the premultiplication of the given state identifier can fit into the representation indicated by S. If it cannot, or if it overflows usize itself, then an error is returned.

usize_to_state_id

Convert the given usize to the chosen state identifier representation. If the given value cannot fit in the chosen representation, then an error is returned.