Struct aho_corasick::MatchError

pub struct MatchError(_);

An error that occurred during an Aho-Corasick search.

An error that occurs during a search is limited to some kind of misconfiguration that resulted in an illegal call. Stated differently, whether an error occurs is not dependent on the specific bytes in the haystack.

Examples of misconfiguration:

Executing a stream or overlapping search on a searcher that was built was something other than MatchKind::Standard semantics.

Requested an anchored or an unanchored search on a searcher that doesn’t support unanchored or anchored searches, respectively.

When the std feature is enabled, this implements the std::error::Error trait.

Implementations

impl MatchError

pub fn new(kind: MatchErrorKind) -> MatchError

Create a new error value with the given kind.

This is a more verbose version of the kind-specific constructors, e.g., MatchError::unsupported_stream.

pub fn kind(&self) -> &MatchErrorKind

Returns a reference to the underlying error kind.

pub fn invalid_input_anchored() -> MatchError

Create a new “invalid anchored search” error. This occurs when the caller requests an anchored search but where anchored searches aren’t supported.

This is the same as calling MatchError::new with a MatchErrorKind::InvalidInputAnchored kind.

pub fn invalid_input_unanchored() -> MatchError

Create a new “invalid unanchored search” error. This occurs when the caller requests an unanchored search but where unanchored searches aren’t supported.

This is the same as calling MatchError::new with a MatchErrorKind::InvalidInputUnanchored kind.

pub fn unsupported_stream(got: MatchKind) -> MatchError

Create a new “unsupported stream search” error. This occurs when the caller requests a stream search while using an Aho-Corasick automaton with a match kind other than MatchKind::Standard.

The match kind given should be the match kind of the automaton. It should never be MatchKind::Standard.

pub fn unsupported_overlapping(got: MatchKind) -> MatchError

Create a new “unsupported overlapping search” error. This occurs when the caller requests an overlapping search while using an Aho-Corasick automaton with a match kind other than MatchKind::Standard.

The match kind given should be the match kind of the automaton. It should never be MatchKind::Standard.

pub fn unsupported_empty() -> MatchError

Create a new “unsupported empty pattern” error. This occurs when the caller requests a search for which matching an automaton that contains an empty pattern string is not supported.