Struct aho_corasick::PatternIDError

pub struct PatternIDError(/* private fields */);

This error occurs when an ID could not be constructed.

This occurs when given an integer exceeding the maximum allowed value.

When the std feature is enabled, this implements the Error trait.

Implementations

impl PatternIDError

pub fn attempted(&self) -> u64

Returns the value that could not be converted to an ID.