aho_corasick

Module util

Modules

alphabet 🔒

buffer 🔒

byte_frequencies 🔒

debug 🔒

error 🔒

int 🔒

This module provides several integer oriented traits for converting between both fixed size integers and integers whose size varies based on the target (like usize).

prefilter 🔒

primitives 🔒

Lower level primitive types that are useful in a variety of circumstances.

remapper 🔒

search 🔒

special 🔒