aho_corasick::util

Module alphabet

Structs

BitSet 🔒

The representation of a byte set. Split out so that we can define a convenient Debug impl for it while keeping “ByteSet” in the output.

ByteClassElementRanges 🔒

An iterator over all elements in an equivalence class expressed as a sequence of contiguous ranges.

ByteClassElements 🔒

An iterator over all elements in a specific equivalence class.

ByteClassIter 🔒

An iterator over each equivalence class.

ByteClassSet 🔒

A partitioning of bytes into equivalence classes.

ByteClasses 🔒

A representation of byte oriented equivalence classes.

ByteSet 🔒

A simple set of bytes that is reasonably cheap to copy and allocation free.