aho_corasick::util::prefilter
Function opposite_ascii_case

pub(crate) fn opposite_ascii_case(b: u8) -> u8

If the given byte is an ASCII letter, then return it in the opposite case. e.g., Given b'A', this returns b'a', and given b'a', this returns b'A'. If a non-ASCII letter is given, then the given byte is returned.