aho_corasick
Module ahocorasick

Structs

AhoCorasick

An automaton for searching multiple strings in linear time.

AhoCorasickBuilder

A builder for configuring an Aho-Corasick automaton.

FindIter

An iterator of non-overlapping matches in a particular haystack.

FindOverlappingIter

An iterator of overlapping matches in a particular haystack.

StreamFindIter

An iterator that reports Aho-Corasick matches in a stream.

Enums

AhoCorasickKind

The type of Aho-Corasick implementation to use in an AhoCorasick searcher.

Traits

AcAutomaton 🔒

A trait that effectively gives us practical dynamic dispatch over anything that impls Automaton, but without needing to add a bunch of bounds to the core Automaton trait. Basically, we provide all of the marker traits that our automatons have, in addition to Debug impls and requiring that there is no borrowed data. Without these, the main AhoCorasick type would not be able to meaningfully impl Debug or the marker traits without also requiring that all impls of Automaton do so, which would be not great.

Functions

enforce_anchored_consistency 🔒

Returns an error if the start state configuration does not support the desired search configuration. See the internal ‘AhoCorasick::start_kind’ field docs for more details.