Module aho_corasick::automaton

Provides Automaton trait for abstracting over Aho-Corasick automata.

The Automaton trait provides a way to write generic code over any Aho-Corasick automaton. It also provides access to lower level APIs that permit walking the state transitions of an Aho-Corasick automaton manually.

Structs

FindIter

An iterator of non-overlapping matches in a particular haystack.

FindOverlappingIter

An iterator of overlapping matches in a particular haystack.

OverlappingState

Represents the current state of an overlapping search.

Prefilter

A prefilter for accelerating a search.

StateID

The identifier of a finite automaton state.

StateIDError

This error occurs when an ID could not be constructed.

StreamFindIter

An iterator that reports matches in a stream.

Enums

Candidate

A candidate is the result of running a prefilter on a haystack at a particular position.

Traits

Automaton

A trait that abstracts over Aho-Corasick automata.