Trait aho_corasick::automaton::Automaton

pub trait Automaton {
    type ID: StateID;
    fn match_kind(&self) -> &MatchKind;

    fn prefilter(&self) -> Option<&dyn Prefilter>;

    fn start_state(&self) -> Self::ID;

    fn is_valid(&self, id: Self::ID) -> bool;

    fn is_match_state(&self, id: Self::ID) -> bool;

    fn get_match(
        &self, 
        id: Self::ID, 
        match_index: usize, 
        end: usize
    ) -> Option<Match>;

    fn match_count(&self, id: Self::ID) -> usize;

    unsafe fn next_state_unchecked(
        &self, 
        current: Self::ID, 
        input: u8
    ) -> Self::ID;

    fn is_match_or_dead_state(&self, id: Self::ID) -> bool { ... }

    unsafe fn next_state_unchecked_no_fail(
        &self, 
        current: Self::ID, 
        input: u8
    ) -> Self::ID { ... }

    fn standard_find_at(
        &self, 
        prestate: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }

    fn standard_find_at_imp(
        &self, 
        prestate: &mut PrefilterState, 
        prefilter: Option<&dyn Prefilter>, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }

    fn leftmost_find_at(
        &self, 
        prestate: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }

    fn leftmost_find_at_imp(
        &self, 
        prestate: &mut PrefilterState, 
        prefilter: Option<&dyn Prefilter>, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }

    fn overlapping_find_at(
        &self, 
        prestate: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID, 
        match_index: &mut usize
    ) -> Option<Match> { ... }

    fn earliest_find_at(
        &self, 
        prestate: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }

    fn find_at(
        &self, 
        prestate: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize, 
        state_id: &mut Self::ID
    ) -> Option<Match> { ... }
}

A trait describing the interface of an Aho-Corasick finite state machine.

Every automaton has exactly one fail state, one dead state and exactly one start state. Generally, these correspond to the first, second and third states, respectively. The failure state is always treated as a sentinel. That is, no correct Aho-Corasick automaton will ever transition into the fail state. The dead state, however, can be transitioned into, but only when leftmost-first or leftmost-longest match semantics are enabled and only when at least one match has been observed.

Every automaton also has one or more match states, such that Automaton::is_match_state_unchecked(id) returns true if and only if id corresponds to a match state.

Associated Types

type ID: StateID

The representation used for state identifiers in this automaton.

Typically, this is one of u8, u16, u32, u64 or usize.

Loading content...
Required methods

fn match_kind(&self) -> &MatchKind

The type of matching that should be done.

fn prefilter(&self) -> Option<&dyn Prefilter>

An optional prefilter for quickly skipping to the next candidate match. A prefilter must report at least every match, although it may report positions that do not correspond to a match. That is, it must not allow false negatives, but can allow false positives.

Currently, a prefilter only runs when the automaton is in the start state. That is, the position reported by a prefilter should always correspond to the start of a potential match.

fn start_state(&self) -> Self::ID

Return the identifier of this automaton's start state.

fn is_valid(&self, id: Self::ID) -> bool

Returns true if and only if the given state identifier refers to a valid state.

fn is_match_state(&self, id: Self::ID) -> bool

Returns true if and only if the given identifier corresponds to a match state.

The state ID given must be valid, or else implementors may panic.

fn get_match(
&self,
id: Self::ID,
match_index: usize,
end: usize
) -> Option<Match>

If the given state is a match state, return the match corresponding to the given match index. end must be the ending position of the detected match. If no match exists or if match_index exceeds the number of matches in this state, then None is returned.

The state ID given must be valid, or else implementors may panic.

If the given state ID is correct and if the match_index is less than the number of matches for that state, then this is guaranteed to return a match.

fn match_count(&self, id: Self::ID) -> usize

Returns the number of matches for the given state. If the given state is not a match state, then this returns 0.

The state ID given must be valid, or else implementors must panic.

unsafe fn next_state_unchecked(&self, current: Self::ID, input: u8) -> Self::ID

Given the current state that this automaton is in and the next input byte, this method returns the identifier of the next state. The identifier returned must always be valid and may never correspond to the fail state. The returned identifier may, however, point to the dead state.

This is not safe so that implementors may look up the next state without memory safety checks such as bounds checks. As such, callers must ensure that the given identifier corresponds to a valid automaton state. Implementors must, in turn, ensure that this routine is safe for all valid state identifiers and for all possible u8 values.

Loading content...
Provided methods

fn is_match_or_dead_state(&self, id: Self::ID) -> bool

Returns true if and only if the given identifier corresponds to a state that is either the dead state or a match state.

Depending on the implementation of the automaton, this routine can be used to save a branch in the core matching loop. Nevertheless, is_match_state(id) || id == dead_id() is always a valid implementation. Indeed, this is the default implementation.

The state ID given must be valid, or else implementors may panic.

unsafe fn next_state_unchecked_no_fail(
&self,
current: Self::ID,
input: u8
) -> Self::ID

Like next_state_unchecked, but debug_asserts that the underlying implementation never returns a fail_id() for the next state.

fn standard_find_at(
&self,
prestate: &mut PrefilterState,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

Execute a search using standard match semantics.

This can be used even when the automaton was constructed with leftmost match semantics when you want to find the earliest possible match. This can also be used as part of an overlapping search implementation.

N.B. This does not report a match if state_id is given as a matching state. As such, this should not be used directly.

fn standard_find_at_imp(
&self,
prestate: &mut PrefilterState,
prefilter: Option<&dyn Prefilter>,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

fn leftmost_find_at(
&self,
prestate: &mut PrefilterState,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

Execute a search using leftmost (either first or longest) match semantics.

The principle difference between searching with standard semantics and searching with leftmost semantics is that leftmost searching will continue searching even after a match has been found. Once a match is found, the search does not stop until either the haystack has been exhausted or a dead state is observed in the automaton. (Dead states only exist in automatons constructed with leftmost semantics.) That is, we rely on the construction of the automaton to tell us when to quit.

fn leftmost_find_at_imp(
&self,
prestate: &mut PrefilterState,
prefilter: Option<&dyn Prefilter>,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

fn overlapping_find_at(
&self,
prestate: &mut PrefilterState,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID,
match_index: &mut usize
) -> Option<Match>

Execute an overlapping search.

When executing an overlapping match, the previous state ID in addition to the previous match index should be given. If there are more matches at the given state, then the match is reported and the given index is incremented.

fn earliest_find_at(
&self,
prestate: &mut PrefilterState,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

Return the earliest match found. This returns as soon as we know that we have a match. As such, this does not necessarily correspond to the leftmost starting match, but rather, the leftmost position at which a match ends.

fn find_at(
&self,
prestate: &mut PrefilterState,
haystack: &[u8],
at: usize,
state_id: &mut Self::ID
) -> Option<Match>

A convenience function for finding the next match according to the match semantics of this automaton. For standard match semantics, this finds the earliest match. Otherwise, the leftmost match is found.

Loading content...