Module aho_corasick::dfa

Provides direct access to a DFA implementation of Aho-Corasick.

This is a low-level API that generally only needs to be used in niche circumstances. When possible, prefer using AhoCorasick instead of a DFA directly. Using an DFA directly is typically only necessary when one needs access to the Automaton trait implementation.

Structs

Builder

A builder for configuring an Aho-Corasick DFA.

DFA

A DFA implementation of Aho-Corasick.