Module aho_corasick::dfa

Structs

Builder
A builder for configuring the determinization of an NFA into a DFA.

ByteClass

Premultiplied

PremultipliedByteClass

Repr

Standard

Enums

DFA

Functions

nfa_next_state_memoized
This returns the next NFA transition (including resolving failure transitions), except once it sees a state id less than the id of the DFA state that is currently being populated, then we no longer need to follow failure transitions and can instead query the pre-computed state id from the DFA itself.