Struct aho_corasick::dfa::Repr

pub struct Repr<S> {

    match_kind: MatchKind,
    anchored: bool,
    premultiplied: bool,
    start_id: S,
    max_pattern_len: usize,
    pattern_count: usize,
    state_count: usize,
    max_match: S,
    heap_bytes: usize,
    prefilter: Option<PrefilterObj>,
    byte_classes: ByteClasses,
    trans: Vec<S>,
    matches: Vec<Vec<(usize, usize)>>,

}

Fields
match_kind: MatchKindanchored: boolpremultiplied: boolstart_id: Smax_pattern_len: usize

The length, in bytes, of the longest pattern in this automaton. This information is useful for keeping correct buffer sizes when searching on streams.

pattern_count: usize

The total number of patterns added to this automaton. This includes patterns that may never match.

state_count: usizemax_match: Sheap_bytes: usize

The number of bytes of heap used by this NFA’s transition table.

prefilter: Option<PrefilterObj>

A prefilter for quickly detecting candidate matchs, if pertinent.

byte_classes: ByteClassestrans: Vec<S>matches: Vec<Vec<(usize, usize)>>
Implementations

impl<S: StateID> Repr<S>

fn alphabet_len(&self) -> usize

Returns the total alphabet size for this DFA.

If byte classes are enabled, then this corresponds to the number of equivalence classes. If they are disabled, then this is always 256.

fn is_match_state(&self, id: S) -> bool

Returns true only if the given state is a match state.

fn is_match_or_dead_state(&self, id: S) -> bool

Returns true only if the given state is either a dead state or a match state.

fn get_match(&self, id: S, match_index: usize, end: usize) -> Option<Match>

Get the ith match for the given state, where the end position of a match was found at end.

Panics

The caller must ensure that the given state identifier is valid, otherwise this may panic. The match_index need not be valid. That is, if the given state has no matches then this returns None.

fn match_count(&self, id: S) -> usize

Return the total number of matches for the given state.

Panics

The caller must ensure that the given identifier is valid, or else this panics.

fn next_state(&self, from: S, byte: u8) -> S

Get the next state given from as the current state and byte as the current input byte.

fn set_next_state(&mut self, from: S, byte: u8, to: S)

Set the byte transition for the from state to point to to.

fn swap_states(&mut self, id1: S, id2: S)

Swap the given states in place.

fn shuffle_match_states(&mut self)

This routine shuffles all match states in this DFA to the beginning of the DFA such that every non-match state appears after every match state. (With one exception: the special fail and dead states remain as the first two states.)

The purpose of doing this shuffling is to avoid an extra conditional in the search loop, and in particular, detecting whether a state is a match or not does not need to access any memory.

This updates self.max_match to point to the last matching state as well as self.start if the starting state was moved.

fn premultiply(&mut self) -> Result<(), Error>

fn calculate_size(&mut self)

Computes the total amount of heap used by this NFA in bytes.