Enum aho_corasick::StartKind

pub enum StartKind {
    Both,
    Unanchored,
    Anchored,
}

The kind of anchored starting configurations to support in an Aho-Corasick searcher.

Depending on which searcher is used internally by AhoCorasick, supporting both unanchored and anchored searches can be quite costly. For this reason, AhoCorasickBuilder::start_kind can be used to configure whether your searcher supports unanchored, anchored or both kinds of searches.

This searcher configuration knob works in concert with the search time configuration Input::anchored. Namely, if one requests an unsupported anchored mode, then the search will either panic or return an error, depending on whether youโre using infallible or fallibe APIs, respectively.

AhoCorasick by default only supports unanchored searches.

Variants

Both

Support both anchored and unanchored searches.

Unanchored

Support only unanchored searches. Requesting an anchored search will return an error in fallible APIs and panic in infallible APIs.

Anchored

Support only anchored searches. Requesting an unanchored search will return an error in fallible APIs and panic in infallible APIs.