aho_corasick::packed

Module api

Structs

Builder

A builder for constructing a packed searcher from a collection of patterns.

Config

The configuration for a packed multiple pattern searcher.

FindIter

An iterator over non-overlapping matches from a packed searcher.

Searcher

A packed searcher for quickly finding occurrences of multiple patterns.

Enums

ForceAlgorithm 🔒

An internal option for forcing the use of a particular packed algorithm.

MatchKind

A knob for controlling the match semantics of a packed multiple string searcher.

SearchKind 🔒

Constants

PATTERN_LIMIT 🔒

This is a limit placed on the total number of patterns we’re willing to try and match at once. As more sophisticated algorithms are added, this number may be increased.