Struct aho_corasick::packed::pattern::Patterns

pub struct Patterns {
    kind: MatchKind,
    by_id: Vec<Vec<u8>>,
    order: Vec<u16>,
    minimum_len: usize,
    max_pattern_id: u16,
    total_pattern_bytes: usize,
}

A non-empty collection of non-empty patterns to search for.

This collection of patterns is what is passed around to both execute searches and to construct the searchers themselves. Namely, this permits searches to avoid copying all of the patterns, and allows us to keep only one copy throughout all packed searchers.

Note that this collection is not a set. The same pattern can appear more than once.

Fields
kind: MatchKind

The match semantics supported by this collection of patterns.

The match semantics determines the order of the iterator over patterns. For leftmost-first, patterns are provided in the same order as were provided by the caller. For leftmost-longest, patterns are provided in descending order of length, with ties broken by the order in which they were provided by the caller.

by_id: Vec<Vec<u8>>

The collection of patterns, indexed by their identifier.

order: Vec<u16>

The order of patterns defined for iteration, given by pattern identifiers. The order of by_id and order is always the same for leftmost-first semantics, but may be different for leftmost-longest semantics.

minimum_len: usize

The length of the smallest pattern, in bytes.

max_pattern_id: u16

The largest pattern identifier. This should always be equivalent to the number of patterns minus one in this collection.

total_pattern_bytes: usize

The total number of pattern bytes across the entire collection. This is used for reporting total heap usage in constant time.

Implementations

impl Patterns

pub fn new() -> Patterns

Create a new collection of patterns for the given match semantics. The ID of each pattern is the index of the pattern at which it occurs in the by_id slice.

If any of the patterns in the slice given are empty, then this panics. Similarly, if the number of patterns given is zero, then this also panics.

pub fn add(&mut self, bytes: &[u8])

Add a pattern to this collection.

This panics if the pattern given is empty.

pub fn set_match_kind(&mut self, kind: MatchKind)

Set the match kind semantics for this collection of patterns.

If the kind is not set, then the default is leftmost-first.

pub fn len(&self) -> usize

Return the number of patterns in this collection.

This is guaranteed to be greater than zero.

pub fn is_empty(&self) -> bool

Returns true if and only if this collection of patterns is empty.

pub fn heap_bytes(&self) -> usize

Returns the approximate total amount of heap used by these patterns, in units of bytes.

pub fn reset(&mut self)

Clears all heap memory associated with this collection of patterns and resets all state such that it is a valid empty collection.

pub fn max_pattern_id(&self) -> u16

Return the maximum pattern identifier in this collection. This can be useful in searchers for ensuring that the collection of patterns they are provided at search time and at build time have the same size.

pub fn minimum_len(&self) -> usize

Returns the length, in bytes, of the smallest pattern.

This is guaranteed to be at least one.

pub fn match_kind(&self) -> &MatchKind

Returns the match semantics used by these patterns.

pub fn get(&self, id: u16) -> Pattern<'_>

Return the pattern with the given identifier. If such a pattern does not exist, then this panics.

pub unsafe fn get_unchecked(&self, id: u16) -> Pattern<'_>

Return the pattern with the given identifier without performing bounds checks.

Safety

Callers must ensure that a pattern with the given identifier exists before using this method.

pub fn iter(&self) -> PatternIter<'_>

Return an iterator over all the patterns in this collection, in the order in which they should be matched.

Specifically, in a naive multi-pattern matcher, the following is guaranteed to satisfy the match semantics of this collection of patterns:

for i in 0..haystack.len():
  for p in patterns.iter():
    if haystack[i..].starts_with(p.bytes()):
      return Match(p.id(), i, i + p.bytes().len())

Namely, among the patterns in a collection, if they are matched in the order provided by this iterator, then the result is guaranteed to satisfy the correct match semantics. (Either leftmost-first or leftmost-longest.)