Struct aho_corasick::packed::Builder

pub struct Builder { /* private fields */ }

A builder for constructing a packed searcher from a collection of patterns.

Example

This example shows how to use a builder to construct a searcher. By default, leftmost-first match semantics are used.

use aho_corasick::{packed::{Builder, MatchKind}, PatternID};

let searcher = Builder::new()
    .add("foobar")
    .add("foo")
    .build()?;
let matches: Vec<PatternID> = searcher
    .find_iter("foobar")
    .map(|mat| mat.pattern())
    .collect();
assert_eq!(vec![PatternID::ZERO], matches);

Implementations

impl Builder

pub fn new() -> Builder

Create a new builder for constructing a multi-pattern searcher. This constructor uses the default configuration.

pub fn build(&self) -> Option<Searcher>

Build a searcher from the patterns added to this builder so far.

pub fn add<P: AsRef<[u8]>>(&mut self, pattern: P) -> &mut Builder

Add the given pattern to this set to match.

The order in which patterns are added is significant. Namely, when using leftmost-first match semantics, then when multiple patterns can match at a particular location, the pattern that was added first is used as the match.

If the number of patterns added exceeds the amount supported by packed searchers, then the builder will stop accumulating patterns and render itself inert. At this point, constructing a searcher will always return None.

pub fn extend<I, P>(&mut self, patterns: I) -> &mut Builderwhere I: IntoIterator<Item = P>, P: AsRef<[u8]>,

Add the given iterator of patterns to this set to match.

The iterator must yield elements that can be converted into a &[u8].

The order in which patterns are added is significant. Namely, when using leftmost-first match semantics, then when multiple patterns can match at a particular location, the pattern that was added first is used as the match.

If the number of patterns added exceeds the amount supported by packed searchers, then the builder will stop accumulating patterns and render itself inert. At this point, constructing a searcher will always return None.

pub fn len(&self) -> usize

Returns the number of patterns added to this builder.

pub fn minimum_len(&self) -> usize

Returns the length, in bytes, of the shortest pattern added.