Module aho_corasick::prefilter

Structs

PrefilterObj
A convenience object for representing any type that implements Prefilter and is cloneable.

PrefilterState
PrefilterState tracks state associated with the effectiveness of a prefilter. It is used to track how many bytes, on average, are skipped by the prefilter. If this average dips below a certain threshold over time, then the state renders the prefilter inert and stops using it.

StartBytesBuilder
A builder for construction a starting byte prefilter.

StartBytesOne
A prefilter for scanning for a single starting byte.

StartBytesThree
A prefilter for scanning for three starting bytes.

StartBytesTwo
A prefilter for scanning for two starting bytes.

Traits

Prefilter
A prefilter describes the behavior of fast literal scanners for quickly skipping past bytes in the haystack that we know cannot possibly participate in a match.