Trait aho_corasick::prefilter::Prefilter

pub trait Prefilter: Send + Sync + RefUnwindSafe + UnwindSafe + Debug {
    fn next_candidate(&self, haystack: &[u8], at: usize) -> Option<usize>;

    fn clone_prefilter(&self) -> Box<dyn Prefilter>;
}

A prefilter describes the behavior of fast literal scanners for quickly skipping past bytes in the haystack that we know cannot possibly participate in a match.

Required methods

fn next_candidate(&self, haystack: &[u8], at: usize) -> Option<usize>

Returns the next possible match candidate. This may yield false positives, so callers must "confirm" a match starting at the position returned. This, however, must never produce false negatives. That is, this must, at minimum, return the starting position of the next match in the given haystack after or at the given position.

fn clone_prefilter(&self) -> Box<dyn Prefilter>

A method for cloning a prefilter, to work-around the fact that Clone is not object-safe.

Loading content...