Trait aho_corasick::prefilter::Prefilter

pub trait Prefilter: Send + Sync + RefUnwindSafe + UnwindSafe + Debug {
    fn next_candidate(
        &self, 
        state: &mut PrefilterState, 
        haystack: &[u8], 
        at: usize
    ) -> Candidate;

    fn clone_prefilter(&self) -> Box<dyn Prefilter>;

    fn heap_bytes(&self) -> usize;

    fn reports_false_positives(&self) -> bool { ... }

    fn looks_for_non_start_of_match(&self) -> bool { ... }
}

A prefilter describes the behavior of fast literal scanners for quickly skipping past bytes in the haystack that we know cannot possibly participate in a match.

Required methods

fn next_candidate(
&self,
state: &mut PrefilterState,
haystack: &[u8],
at: usize
) -> Candidate

Returns the next possible match candidate. This may yield false positives, so callers must confirm a match starting at the position returned. This, however, must never produce false negatives. That is, this must, at minimum, return the starting position of the next match in the given haystack after or at the given position.

fn clone_prefilter(&self) -> Box<dyn Prefilter>

A method for cloning a prefilter, to work-around the fact that Clone is not object-safe.

fn heap_bytes(&self) -> usize

Returns the approximate total amount of heap used by this prefilter, in units of bytes.

Provided methods

fn reports_false_positives(&self) -> bool

Returns true if and only if this prefilter never returns false positives. This is useful for completely avoiding the automaton when the prefilter can quickly confirm its own matches.

By default, this returns true, which is conservative; it is always correct to return true. Returning false here and reporting a false positive will result in incorrect searches.

fn looks_for_non_start_of_match(&self) -> bool

Returns true if and only if this prefilter may look for a non-starting position of a match.

This is useful in a streaming context where prefilters that don’t look for a starting position of a match can be quite difficult to deal with.

This returns false by default.