Struct aho_corasick::Match

pub struct Match {
    pub pati: usize,
    pub start: usize,
    pub end: usize,
}

Records a match in the search text.

Fields

pati

The pattern index.

This corresponds to the ordering in which the matched pattern was added to the automaton, starting at 0.

start

The starting byte offset of the match in the search text.

end

The ending byte offset of the match in the search text.

(This can be re-captiulated with pati and adding the pattern's length to start, but it is convenient to have it here.)