Struct aho_corasick::Match

pub struct Match { /* fields omitted */ }

A representation of a match reported by an Aho-Corasick automaton.

A match has two essential pieces of information: the identifier of the pattern that matched, along with the start and end offsets of the match in the haystack.

Examples

Basic usage:

use aho_corasick::AhoCorasick;

let ac = AhoCorasick::new(&[
    "foo", "bar", "baz",
]);
let mat = ac.find("xxx bar xxx").expect("should have a match");
assert_eq!(1, mat.pattern());
assert_eq!(4, mat.start());
assert_eq!(7, mat.end());

Methods

impl Match

pub fn pattern(&self) -> usize

Returns the identifier of the pattern that matched.

The identifier of a pattern is derived from the position in which it was originally inserted into the corresponding automaton. The first pattern has identifier 0, and each subsequent pattern is 1, 2 and so on.

pub fn start(&self) -> usize

The starting position of the match.

pub fn end(&self) -> usize

The ending position of the match.

pub fn is_empty(&self) -> bool

Returns true if and only if this match is empty. That is, when start() == end().

An empty match can only be returned when the empty string was among the patterns used to build the Aho-Corasick automaton.