aho_corasick
Struct Match

pub struct Match { /* private fields */ }

A representation of a match reported by an Aho-Corasick searcher.

A match has two essential pieces of information: the PatternID that matches, and the Span of the match in a haystack.

The pattern is identified by an ID, which corresponds to its position (starting from 0) relative to other patterns used to construct the corresponding searcher. If only a single pattern is provided, then all matches are guaranteed to have a pattern ID of 0.

Every match reported by a searcher guarantees that its span has its start offset as less than or equal to its end offset.

Implementations

impl Match

pub fn new<S: Into<Span>>(pattern: PatternID, span: S) -> Match

Create a new match from a pattern ID and a span.

This constructor is generic over how a span is provided. While a Span may be given directly, one may also provide a std::ops::Range<usize>.

Panics

This panics if end < start.

Example

This shows how to create a match for the first pattern in an Aho-Corasick searcher using convenient range syntax.

use aho_corasick::{Match, PatternID};

let m = Match::new(PatternID::ZERO, 5..10);
assert_eq!(0, m.pattern().as_usize());
assert_eq!(5, m.start());
assert_eq!(10, m.end());

pub fn must<S: Into<Span>>(pattern: usize, span: S) -> Match

Create a new match from a pattern ID and a byte offset span.

This constructor is generic over how a span is provided. While a Span may be given directly, one may also provide a std::ops::Range<usize>.

This is like Match::new, but accepts a usize instead of a PatternID. This panics if the given usize is not representable as a PatternID.

Panics

This panics if end < start or if pattern > PatternID::MAX.

Example

This shows how to create a match for the third pattern in an Aho-Corasick searcher using convenient range syntax.

use aho_corasick::Match;

let m = Match::must(3, 5..10);
assert_eq!(3, m.pattern().as_usize());
assert_eq!(5, m.start());
assert_eq!(10, m.end());

pub fn pattern(&self) -> PatternID

Returns the ID of the pattern that matched.

The ID of a pattern is derived from the position in which it was originally inserted into the corresponding searcher. The first pattern has identifier 0, and each subsequent pattern is 1, 2 and so on.

pub fn start(&self) -> usize

The starting position of the match.

This is a convenience routine for Match::span().start.

pub fn end(&self) -> usize

The ending position of the match.

This is a convenience routine for Match::span().end.

pub fn range(&self) -> Range<usize>

Returns the match span as a range.

This is a convenience routine for Match::span().range().

pub fn span(&self) -> Span

Returns the span for this match.

pub fn is_empty(&self) -> bool

Returns true when the span in this match is empty.

An empty match can only be returned when empty pattern is in the Aho-Corasick searcher.

pub fn len(&self) -> usize

Returns the length of this match.

This returns 0 in precisely the cases that is_empty returns true.

pub fn offset(&self, offset: usize) -> Match

Returns a new match with offset added to its spanโs start and end values.