Struct aho_corasick::Span

pub struct Span {
    pub start: usize,
    pub end: usize,
}

A representation of a range in a haystack.

A span corresponds to the starting and ending byte offsets of a contiguous region of bytes. The starting offset is inclusive while the ending offset is exclusive. That is, a span is a half-open interval.

A span is used to report the offsets of a match, but it is also used to convey which region of a haystack should be searched via routines like Input::span.

This is basically equivalent to a std::ops::Range<usize>, except this type implements Copy which makes it more ergonomic to use in the context of this crate. Indeed, Span exists only because Range<usize> does not implement Copy. Like a range, this implements Index for [u8] and str, and IndexMut for [u8]. For convenience, this also impls From<Range>, which means things like Span::from(5..10) work.

There are no constraints on the values of a span. It is, for example, legal to create a span where start > end.

Fields
start: usize

The start offset of the span, inclusive.

end: usize

The end offset of the span, exclusive.

Implementations

impl Span

pub fn range(&self) -> Range<usize>

Returns this span as a range.

pub fn is_empty(&self) -> bool

Returns true when this span is empty. That is, when start >= end.

pub fn len(&self) -> usize

Returns the length of this span.

This returns 0 in precisely the cases that is_empty returns true.

pub fn contains(&self, offset: usize) -> bool

Returns true when the given offset is contained within this span.

Note that an empty span contains no offsets and will always return false.

pub fn offset(&self, offset: usize) -> Span

Returns a new span with offset added to this spanโs start and end values.