aho_corasick

Module util

Modules

alphabet ğŸ”’

buffer ğŸ”’

byte_frequencies ğŸ”’

debug ğŸ”’

error ğŸ”’

int ğŸ”’

This module provides several integer oriented traits for converting between both fixed size integers and integers whose size varies based on the target (like usize).

prefilter ğŸ”’

primitives ğŸ”’

Lower level primitive types that are useful in a variety of circumstances.

remapper ğŸ”’

search ğŸ”’

special ğŸ”’